Fixes #343. Improved the code.
