--- desktop/ui/documentLoaderOutput.html
@@ -1,220 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-<link rel="stylesheet" type="text/css" href=C:\Users\claudiap\git\DataCleaner\desktop\ui\src\main\resources\documentation.css />
-</head>
-<body class="body">
-	<div class="navbar">
-		<a class="navbar-brand"> DataCleaner </a>
-	</div>
-	<div class="documentationBody">
-		<h1 class="title">
-			<image src=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAgCAYAAABzenr0AAABRUlEQVR42mNgGAyg+rza/4HAow7AcMCK+4X/QQBED4gDLrzd/P/6h31AehOKwkMvZv//8uvt/++/P/8/+2YdQXGyHFBzTuP/519v/ndfcQLSb8F8dA0Trnn+//v3L9HiWC3DlQtm34r+jwxAfJjm+beT/197v/f/ux+PwXKExMlywOEXc/9vfdwGVgyiD72YA9f89feH/8vvFfxf/7AGxSJc4mQ54NX3e/97r7qCFYPoV9/uwjWD4vf7709gRyJbhEucLAfQLcUPSgcYGxv/pwTfuXMHKya6Lhh1AMiQlStX/v/8+fP/T58+/V+wYAHccFziNAuB5ubm/zdu3MCwBJc4VR3Q39////jx4/8dHR1RLMAlTnUHgILZwcEBwwJ0cVwWkmQpJYmQZg4AAWwWoosP3xAYdQAhw6lm4agDRh0w6oDBBAA6bKn5UjgQFgAAAABJRU5ErkJggg== alt="Image"> Pattern finder 
-		</h1>
-
-		<h4>The Pattern Finder will inspect your String values and generate and match string patterns that suit your data.
-It can be used for a lot of purposes but is excellent for verifying or getting ideas about the format of the string-values in a column.</h4>
-		<div class="details">
-			<ol>
-				<li><h3>Execution details</h3></li>
-				<ul>
-					<li>
-						<dl>
-							<dt>Distributed execution possible</dt>
-							<dd>Yes</dd>
-						</dl>
-					</li>
-				</ul>
-				<li><h3>Configured Properties</h3></li>
-				<div class="propertiesdetails">
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Column  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">InputColumn</dd>
- 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Group column  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">InputColumn</dd>
-							<dt class="col">Description</dt>
-							<dd class="colvalue">Optional column to group patterns by</dd>
-							 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Discriminate text case  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">Boolean</dd>
-							<dt class="col">Description</dt>
-							<dd class="colvalue">Separate text tokens based on case</dd>
-							 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Discriminate negative numbers  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">Boolean</dd>
-							<dt class="col">Description</dt>
-							<dd class="colvalue">Separate number tokens based on negativity</dd>
-							 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Discriminate decimals  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">Boolean</dd>
-							<dt class="col">Description</dt>
-							<dd class="colvalue">Separate number tokens for decimals</dd>
-							 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Enable mixed tokens  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">Boolean</dd>
-							<dt class="col">Description</dt>
-							<dd class="colvalue">Use '?'-tokens for mixed text and numbers</dd>
-							 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Ignore repeated spaces  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">Boolean</dd>
-							<dt class="col">Description</dt>
-							<dd class="colvalue">Ignore whitespace differences</dd>
-							 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Upper case patterns expand in size  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">boolean</dd>
-							<dt class="col">Description</dt>
-							<dd class="colvalue">Auto-adjust/expand uppercase text tokens</dd>
-							 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Lower case patterns expand in size  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">boolean</dd>
-							<dt class="col">Description</dt>
-							<dd class="colvalue">Auto-adjust/expand lowercase text tokens</dd>
-							 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Predefined token name  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">String</dd>
- 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Predefined token regexes  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">String[]</dd>
- 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Decimal separator  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">Character</dd>
- 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Thousands separator  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">Character</dd>
- 						</dl>
-					</section>
-					<hr>
-					<section>
-						<dl class="property">
-							<dt class="col">Property</dt>
-							<dd class="name">
-								Minus sign  <span
-									class="required">Required</span> 
-							</dd>
-							<dt class="col">Type</dt>
-							<dd class="colvalue">Character</dd>
- 						</dl>
-					</section>
-					<hr>
-				</div>
-			</ol>
-		</div>
-	</div>
-</body>
-</html>
