+++ desktop/ui/src/main/resources/documentation.css
@@ -40,17 +40,14 @@
 	margin: 10px;
 }
 
-.propertiesdetails .property {
+.propertiesdetails hr{
+	border-width: 1px; 
+/* 	border-color: #5591e2; */
+} 
 
-}
 
-.propertiesdetails .col {
-	font-weight: bold;
+.propertiesdetails h4 {
 	margin-bottom: 10px;
-}
-
-.propertiesdetails .colvalue {
-	
 }
 
 .propertiesdetails .name {
@@ -62,4 +59,5 @@
 	color: white;
 	padding: 3px;
 	margin-left: 10px;
+	font-family: 'Arial'; 
 }
